Day_8_Automated_System_Health_Monitor/requirements.txt
psutil==7.1.2
python-dotenv==1.2.1